refactor(node-api): modularize gatsby-node page creators
